test(EuiDataGridCell): move EuiDataGrid column options jest tests to a separate file
